Bugfixes and new functions in TuflowModel class.
Bugfix 'main' type and added getFileNamesFromModelFile method.
Request for file of type 'main' file was throwing an error because it
was still trying to use the old file_parts setup. TuflowLoader now
records a reference in TuflowModel with the hash of the main file
(the file that is loaded and calls all the other e.g. tcf).

New function adds a simple way to access all the files in a particular
MODEL file.

Bugfix TuflowModel getModelFilesAllTypes.
Wasn't passing modelfile_type key as a list.

Added getNameFromModelFile function to TuflowModel.
TuflowLoader now adds a hex_hash variable to each TuflowModelFile
created. This can be used as a reference to get any details needed
from the TuflowModel.file_parts dictionary.

New function uses this to return the filename of the TuflowFilePart.
Bugfix variable name in TuflowModelFile.