Fixed naming bug in TuflowModel.missing_model_files.
TuflowLoader prepended an underscore to the name. This has
now been changed to add missing files directly to the
missing_model_files list in the TuflowModel instance held
by the loader rather than setting it after the load.

Tidied up some unecessary old code and comments in the loader
as well.