Major restructure of tuflow package class setup [API Break!].
Major change to the way that data is accessed from the
TuflowModel class. This has been done to try and stop it being
a "God" class improve future maintainability.

The TuflowModel class no longer contains a file_parts dict
containing all TuflowFileParts. It now only holds the
files dict (a dict containing lists of TuflowModelFile's
separated by category (tcf, tgc, tbc, ecf).

The TuflowModelFile's now contain the TuflowFilePart's that
are referenced in the files. This includes an model files
(tcf, tgc, etc) that they hold. The exception is the main entry
file (currently always a tcf, but could be an ecf in the future).
This is not referenced by an file so is stored in the
self.mainfile member of TuflowModel.

The TuflowModelFile and it's corresponding TuflowFile can be
accessed using the methods in TuflowModel getTMFFromTuflowFile
and getTuflowFileFromTMF so that they are easily exchangable.

This has affected TuflowModel, TuflowModelFile and TuflowLoader.

There is also now an Enum in the tuflow __init__ file for
referencing the different file types. This is done to avoid
code duplication and possible issues and bugs caused by having
to update them independently. E.g. TuflowModel.MODEL is no longer
used to access the constants, ship.tuflow.FILEPART_TYPES.MODEL
should be used instead.

Docstrings and tests have been updated for the new setup.

Hopefully this will make everything easier to maintain and use.

Note: these changes break the public API.