Added interpretation of scenario and event filenames.
Some Tuflow filenames can contain placeholders using a tilde
character (e.g. ~s~ or ~e~) that identify locations that should
be replaced with variables handed in at runtime.

This commit includes a check in the loader for these and converts
them to the actual values during load. It also makes changes to
the TuflowFile getPrintableContents method to write out the
original placeholders when re-writing the file.

Test have been added and updated and all pass.