Bugfix problem with reading/writing FMP Bridges.
It seems as though there can be a comment after the 'USBPR'
part of the head of usbpr bridges. I'm not sure how this can
be done, but it was causing some usbpr bridges to be read as
arch bridges instead. Now fixed. Same situation with RIVER
units which now use the same startswith approach to checking
the keyword.

There was an issue with bridges failing to write when there
was no value for a header variable. This is allowed, but
trying to convert an empty str to a float is not. Now fixed.

There was also some strange behaviour with comment sections
seemingly missing a line. I think this has come up before?
It is now changed and APPEARS to work again.