Added EventSourceData class and ars_dict arg to datafileloader.
EventSourceData is used to store all of the BC Event Source data
and similar arguments specified in the control files. These are
collected in this convenience class for easy access and use. A
method has been added to the TuflowLoader to do this.

An args_dict variable has been added to the datafileloader
loadDataFile() function and some of the individual loader
functions. This is not implemented yet but is a place holder
for use soon for providing EventSourceData arguments to the
DATA files.