Improved TuflowLoader warning setup and failed load bug.
It is possible to sucessfully load a model when some model
files are missing. This is done so that the missing model files
can be checked. It does mean that it is possible to "successfully"
load a possibly corrupted model.

The bug was in the EventSourceData creation in TuflowLoader.

The warnign reporting by the file loaders has now been improved
and should be checked after loading a model, e.g.:
if loader.warnings: print 'Something has gone wrong'