Added no_duplicates options to getModelFiles.
Scenario and event logic allows for the same tuflow control
file to be referenced multiple times. This means that when
returning model files the same file can added multiple times.
The new argument lets the caller specify that only one of the
same file should be returned.

TODO:
This needs considering a bit more carefully. Should each control
file only be loaded once? Probably, but that has impacts on the
way the TuflowFile version of the control file interacts with
the TuflowModelFile version etc.