Added .trd files to model files dictionary.
Previously .trd files were treated in a different way. They
were considered an extensin to the file that contained then
rather than being a seperate file that is accessible by all of
the control files. This was a misunderstanding of the way that
they are used. This new approach treats them more correctly.