[Bugfix] issue with GIS files that contain no extension.
Some GIS filepart types can be entered into the control files
without an extension. This was causing problems when writing
out files or copying files as the software didn't know how to
find them.

TuflowLoader now checks for these and trys to work out what
they are. If it succeeds it will add the extension to the file,
if not it will gracefully deal with them, but won't be able to
copy them.