Improvements to add and update row in AIsisUnit.
These were error prone and didn't work very well. Particularly
the updateRow method.

Changes mean that it is now possible to hand ony the values
that you want to update.

Also added write() function to datcollection. This means that
only this function needs to be called rather than doing all
of the work.