Updated datunits to remove file_order member.
The file_order member is superfluous because the order of the
units in the dat file is kept track of the the DatCollection
class. This variable will only cause problems as time goes on.

Also moved all of the RowDataCollection construction into the
constructors of the IsisUnit type classes so that they can be
set up more easily when creating a blank unit, rather than
just reading from file.