Proper implementation of the InitialConditions unit.
This was only reading in and out as text before. It has now
been fully implemented to load and access data.

Implementing this has meant further changes were required to
the DatCollection, RowDataCollection, and AIsisUnit and sub-
types. These included adding flags for requiring initial
conditions and adding support for deleting rows from the
collection.

None of this have been tested yet. Extensive additional test
coverage is now required in this part of the API.