Added new method to DatCollection.
getRow() added to DatCollection and RowDataCollection
classes. These allow to return a dictionary for all of the
values in a particular row index with key=data_type and
val=value.

getRowByName() added to InitialConditions unit which does
the same but takes a section_label instead of index.

These seems increasingly to be a lot of double method
writing going on between the RowDataCollection and the
DatCollection. Need to check if this is necessary.