Finalised support for BC Event Source data.
Support for loading the BC Event Source, BC Event Name,
and BC Event Text variables has been added. This will
use Text/Name only when Source is not found. They are
added to two dictionaries in the TuflowModel, one for
the scen/evt setup and another containing all referenced.

They are also added to all TuflowFileParts for them to
use as required. Currently the only way that these are
used is in the DataFileObject's (materials, bc database,
1d_xs) to resolver placeholder names with values.