Bugfixes in addUnit code and some other.
There were issues in the correct updating of initial conditions
for some units because of mis-set flags.

Edge case in spill unit data where an easting can be provided
but no northing causing error (very unlikely in most cases).

dataobject set method was calling the add method on the
superclass.