Updated to new tuflow package setup - basics only.
The majority of the new tuflow package design is complete.
It still needs the docstrings updating and some of the utility
functions putting in.
All of the tests need updating re-writing to accomodate.

It is, I think, now unicode complient and on it's way to working
with python 2.7+ and 3.2+.

Still need to update all of the isis package and some others.