Massive fixup of the isis module + docs.
A lot of changes to the way the isis package works.
A lot of API breaks! But big a big improvement.

Still need to do a lot more work.

Lots of additional documentation added. Including:
	- an introduction walkthrough on the basic functionality.
	- module and class specific docs.
	- outline of lots of other stuff.

Including:
	- Added HeadDataItem for head_data.
	- row_data dict to store all RowDataCollection's.
	- A lot of changes to method names.
	- Cleared out a lot of unnecessary or duplicated variables.

Note not all AIsisUnit types are updated to the new version yet.
The docs are slightly ahead of the code in a lot of places.