Removed row_pos from DataObject.
DataObject contained a member called row_pos that was used to
identify the row index of the entries in a collection. This was
an artifact of an *very* old setup and is not needed now that
the class is an iterator and RowDataCollection is improved. It
was breaking the encapsulation of RowDataCollection and making
it harder to maintain.