Removed row_pos from DataObject + docs update.
There was an old variable called row_pos in the DataObject class
that was still being used in a few places. It was used to store
the order of the DataObject's in the RowDataCollection. It wasn't
needed anymore and just added confusion, so it's gone.

Added a has_dummy variable and a setDummyRow method to
RowDataCollection. These should not be implemented by all of the
AUnit's (I've updated them all). It creates a blank row of
data in the RowDataCollection so that it will not cause loads
of errors in FMP when the user doesn't add any. As soo as some
real data is added has_dummy is set to False and the dummy
row is deleted.

Added a guide to contributing a new AUnit type to the API to
the docs.

Added an example script that shows how to create an empty
DatCollection and add some units to it in a few different ways.