Bugfix TuflowLoader + custom methods added to ControlFile.
There was a bug in TuflowLoader buildControlFiles() introduced
by some changes in the way that PartHolder dealt with adding
new files to the list. This was causing the list to be setup in
the wrong order.
Added a new method (append(TuflowPart)) to be used for forcing a
part to be added to the end of the list.

Added customPartSearch() method to ControlFile and TuflowModel.
These allow the user to hand in a callback function to operate
on the PartHolder items.

Changed tpart_type to filepart_type to maintain naming consistency
across the rest of the API.