Fixed regression in loader and more docs updates.
There was a breaking regression in the loader caused by the
line number being returned from the RiverUnit readUnitData()
method. This was causing the loader to run in a loop.

Added dat loader and writer to integration tests to check for
these things now.

Updated some mistakes in the code examples and lots of formatting
and typo errors in the docs.