Added basics of LinkedUnits support.
LinkedUnits is a class in the new fmp/unitgroups.py module. It
stored units that directly relate to a specific AUnit. These are
other units that refer directly to it or are linked by a
JunctionUnit.

Added a method to DatCollection called linkedUnits() for
retrieving a LinkedUnits object relating to a specific AUnit.

Added a test to test_DatCollection.py for linkedUnits method.