Fixed issue with slow loading of model with DatLoader.
Even small (<200 node) models were taking seconds to load. This
was inadvertently caused by a refactor to using addRow in the
RowDataCollection to add a complete row, rather than the addItem
methods. addRow gives better control over adding new data and is,
I think, clearer.

The problem was caused by the addRow methods call to deepcopy.
This is done to make user calls to addRow reversible when something
goes wrong. When loading a model this is unecessary (as any data
corruption is a failed load anyway) and was causing a massive
overhead.

Added a kwarg called 'no_copy' to the addRow, updateRow and
deleteRow methods to avoid calls to deepcopy. This should always
be used in readUnitData calls to addRow to avoid the copy.

Profiling shows that model load for ~150 node model goes from
~3 seconds to around 0.15 seconds on my machine when deepcopy
isn't called!