More work on improved fmp model load times.
Extended the use of no_copy keyword in adding and removing row
data when calling the constructor and the factory. Two other
places where row data corruption has no knock-on affect.

I think that this is hopefully fairly well covered now and
profiling of load times how a significant improvement (down from
an original time of ~2.8s on my machine, for a 200 node model,
to ~0.025s! This was a silly refactoring cockup!