[Bugfix] Node label formatting for river and junction.
The spill and lateral labels were being incorrectly read in
in the River unit. The output formatting (in getData) has
been updated to fix this. The easting/northing formatting
was also slightly off and has been corrected.

All labels were being incorrectly separated in the
JunctionUnit. They are now separated on every 12th Char to
avoid these issues.

The test for the RiverUnit have been updated to better
identify any similar issue in the future.

A test_juncitonunit module has been added to the tests to
test the basic input/output code for the junction unit.
More is definitely needed here.