[Bugfix] refh urban runoff choices and published_report settings.
The choices for the sewer_rp_2 values were incorrect? Might need
to watch for this as it seems strange that they were so wrong?
Changed 'RUNOFF' choice to 'RETURN'.

The urban refh calculations require the use of the published
report rather than the DLL. FMP will autoenable this on load if
required but we should handle it here to.
When useUrban(True) is used it wass set published_report to
'REPORT'.