Change to the way no_duplicates works for control files.
Previously no_duplicates == True would check that there hadn't
been a previous command match. This doesn't hold true at all for
files where many different files can be called with the same
command.

Now assignes a getter to the FilePart that defaults to checking
the command but can be overridden. This is done for files which
return a combination of the command and the filename.