Add improved error diagnostics on .dat file load.
When an fmp .dat file is loaded it is not uncommon for a it to fail
on malformed file data, with a ValueError.

This change ensures that the start line of the unit and the file
line that it occures at is printed out to help with debugging the
problem.