Add width() method to RiverUnit.
Method to find the width of the river section from the chainage.
Allow the caller to specify as either active only or all data.
If active only the width will be the chainage between the
deactivation markers.

This should probably be generalised and put into a superclass, or
a function somewhere, to be used by other units that contain
width related row data.