update .yml files to include tests for newer python versions.
Adds 3.7 and 3.8 to travis and appveyor testing suites for
integration tests when pushed to github.

Removed 3.3 and travis seems to be failing to download and
verify against it and the test is falsly failing. It's quite
old now and unlikley to cause a lot of issues for us now
anyway.