Fixes to default values for HTBDY and Bridge units.
Hopefully this doesn't leave too many holes in the validation
of the units as they are read in. I think these are acceptable
but they are from old models and appear to have possibly been
updated with a script. FMP accepts them as valid which is
slightly odd but not completely surprising.

Tests will need to be updated to handle these changes, I expect.