Add fix for default values in header and refhunit.
It seems that FMP will allow a lot of misentered, or missing,
values and then read them incorrectly in to the user interface.
This sets up defaults for blank values and just allows the
incorrect ones to be loaded (as long as they type match).

You don't see it very often. I think it's caused by peoples
scripts updating .dat files, so it shouldn't impact on most
dat loads and functionality, but it will allow the data to be
loaded.

Need to add warnings and errors output to the loader so that the
user to check when this sort of thing has happened.