[Bugfix] Identify and handle the dat file label length variable.
In the general settings of the .dat file there is an integer
number, (almost) always set to 12. It turns out that this is the
label length for the units in the dat file.

This change takes the label length and stores it in the
HeaderUnit and the IntitialConditionsUnit so that it can be
tracked and used when reading writing initial conditions.

Should probably also put a check on updating names in units
to ensure that they don't exceed this value.